requirements-local.txt
backports.tarfile==1.2.0
cryptography==44.0.0
dj-database-url==2.3.0
django-admin-interface==0.29.4
django-allauth==65.9.0
django-environ==0.12.0
django-jet==1.0.8
django-notifications-hq==1.8.3
django-recaptcha==4.0.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
importlib-metadata==8.0.0
jaraco.collections==5.1.0
mysqlclient==2.2.7
pip-chill==1.0.3
platformdirs==4.2.2
pyjwt==2.10.1
pymysql==1.1.1
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==8.0
requests==2.32.4
seaborn==0.13.2
setuptools-scm==8.3.1
tenacity==9.0.0
tomli==2.0.1
whitenoise==6.8.2
wrapt==1.17.2